Only use NVIM as pager for MAN pages
NVIM pager didn't really work out well for anything but MAN pages.
It was very slow and even crashed when doing `git log` in repos
with a large history

I guess the MAN plugin wasn't really made to be a general pager.
It was only made to be man pager.